gcgg

+++ Aparelho.h
@@ -24,8 +24,8 @@ public:
     string listProps()const;
     int getid()const;
     void addProp(string nome, shared_ptr<Propriedade> ptr);
-    void aumentaProp(string nome, int val);
-    void diminuiProp(string nome, int val);
+    void aumentaProp(string nome, double val);
+    void diminuiProp(string nome, double val);
     int getPropValue(string nome)const;
     virtual void mudaEstado(const string & nome) = 0;
     virtual ~Aparelho();
